Large number of changes
Changed from mpf_class to mpfr_class everywhere (large number of changes)
Fixed bugs with specifying the ordering attribute in Context::fromFile
splitting now uses sum rather than sum_kahan
Added relative error options to splitting code
Improved splittingVisualiser to show more detail (It now does the simulation and decreasing the radius in separate steps)
Minor fix to all points max flow code

+++ PMC/pmc.cpp
@@ -74,7 +74,7 @@ namespace networkReliability
 		boost::tie(current, end) = boost::edges(graph);
 		for(;current != end; current++)
 		{
-			vertices.push_back(std::make_pair(current->m_source, current->m_target));
+			vertices.push_back(std::make_pair((int)current->m_source, (int)current->m_target));
 		}
 
 		mpfr_class exponentialRate = -mpfr::log(1 - opProbability);
